fixed warnings in parallel
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2680 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/parallel/datadriven/algorithm/AlgorithmAdaBoostSPVectorizedIdentity.cpp
@@ -51,12 +51,12 @@ void AlgorithmAdaBoostSPVectorizedIdentity::alphaSolver(double& lambda, sg::base
 	WMatrix.generateb(classesSP, rhsSP);
 	if (final)
 	{
-		sg::solver::ConjugateGradientsSP myCG(static_cast<float>(this->imax_final), static_cast<float>(this->epsilon_final));
+		sg::solver::ConjugateGradientsSP myCG(this->imax_final, static_cast<float>(this->epsilon_final));
 		myCG.solve(WMatrix, alphaSP, rhsSP, false, false, -1.0);
 	}
 	else
 	{
-		sg::solver::ConjugateGradientsSP myCG(static_cast<float>(this->imax), static_cast<float>(this->epsilon));
+		sg::solver::ConjugateGradientsSP myCG(this->imax, static_cast<float>(this->epsilon));
 		myCG.solve(WMatrix, alphaSP, rhsSP, false, false, -1.0);
 	}
